Backend/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
Babel==2.10.3
billiard==3.6.4.0
celery==5.2.7
celery-flower==1.0.1
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.7
Django==3.2
django-allauth==0.52.0
django-autoslug==1.9.8
django-cors-headers==3.13.0
django-filter==22.1
django-localflavor==3.1
django-phonenumber-field==7.0.0
django-rest-authemail==2.1.4
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-restframework==0.0.1
django-sms==0.6.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
flower==1.2.0
humanize==4.4.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
msgpack==1.0.4
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==21.3
phonenumbers==8.13.1
Pillow==9.2.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
psycopg2==2.9.5
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.9
python-dotenv==0.21.0
python-stdnum==1.17
python3-openid==3.2.0
pytz==2022.5
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
